Extract BatchBuilder for readability.
