+++ library/src/main/java/com/novoda/downloadmanager/BatchBuilder.java
@@ -1,0 +1,7 @@
+package com.novoda.downloadmanager;
+
+public interface BatchBuilder {
+    BatchFileBuilder downloadFrom(String networkAddress);
+
+    Batch build();
+}
