+++ library/src/main/java/com/novoda/downloadmanager/BatchFileBuilder.java
@@ -1,60 +1,11 @@
 package com.novoda.downloadmanager;
 
-import java.io.File;
+public interface BatchFileBuilder {
+    BatchFileBuilder withIdentifier(DownloadFileId downloadFileId);
 
-final class BatchFileBuilder implements InternalBuilder {
+    BatchFileBuilder saveTo(String path);
 
-    private final DownloadBatchId downloadBatchId;
-    private final String networkAddress;
+    BatchFileBuilder saveTo(String path, String fileName);
 
-    private Optional<DownloadFileId> downloadFileId = Optional.absent();
-    private Optional<String> path = Optional.absent();
-
-    private Batch.InternalBuilder parentBuilder;
-
-    BatchFileBuilder(DownloadBatchId downloadBatchId, String networkAddress) {
-        this.downloadBatchId = downloadBatchId;
-        this.networkAddress = networkAddress;
-    }
-
-    @Override
-    public Builder withParentBuilder(Batch.InternalBuilder parentBuilder) {
-        this.parentBuilder = parentBuilder;
-        return this;
-    }
-
-    @Override
-    public Builder withIdentifier(DownloadFileId downloadFileId) {
-        this.downloadFileId = Optional.fromNullable(downloadFileId);
-        return this;
-    }
-
-    @Override
-    public Builder saveTo(String path) {
-        String networkAddressDerivedFileName = FileNameExtractor.extractFrom(networkAddress);
-        return saveTo(path, networkAddressDerivedFileName);
-    }
-
-    @Override
-    public Builder saveTo(String path, String fileName) {
-        if (path != null && fileName != null) {
-            this.path = Optional.of(path + fileName);
-        }
-
-        return this;
-    }
-
-    @Override
-    public Batch.Builder apply() {
-        String networkAddressDerivedFileName = FileNameExtractor.extractFrom(networkAddress);
-        String pathPrependedWithBatchId = prependBatchIdTo(path.or(networkAddressDerivedFileName), downloadBatchId);
-
-        parentBuilder.withFile(new BatchFile(networkAddress, downloadFileId, pathPrependedWithBatchId));
-        return parentBuilder;
-    }
-
-    private static String prependBatchIdTo(String filePath, DownloadBatchId downloadBatchId) {
-        return downloadBatchId.rawId() + File.separatorChar + filePath;
-    }
-
+    BatchBuilder apply();
 }
