--- library/src/main/java/com/novoda/downloadmanager/Builder.java
@@ -1,11 +1,0 @@
-package com.novoda.downloadmanager;
-
-public interface Builder {
-    Builder withIdentifier(DownloadFileId downloadFileId);
-
-    Builder saveTo(String path);
-
-    Builder saveTo(String path, String fileName);
-
-    Batch.Builder apply();
-}
