+++ library/src/main/java/com/novoda/downloadmanager/InternalBatchBuilder.java
@@ -1,0 +1,5 @@
+package com.novoda.downloadmanager;
+
+interface InternalBatchBuilder extends BatchBuilder {
+    void withFile(BatchFile batchFile);
+}
