+++ library/src/main/java/com/novoda/downloadmanager/InternalBatchFileBuilder.java
@@ -1,0 +1,5 @@
+package com.novoda.downloadmanager;
+
+interface InternalBatchFileBuilder extends BatchFileBuilder {
+    BatchFileBuilder withParentBuilder(InternalBatchBuilder parentBuilder);
+}
