+++ library/src/test/java/com/novoda/downloadmanager/BatchBuilderTest.java
@@ -11,7 +11,7 @@ public class BatchBuilderTest {
     private static final DownloadBatchId DOWNLOAD_BATCH_ID = DownloadBatchIdCreator.createSanitizedFrom("download_batch_id");
     private static final String DOWNLOAD_BATCH_TITLE = "download_batch_title";
     private static final DownloadFileId DOWNLOAD_FILE_ID = new LiteDownloadFileId("download_file_id");
-    private static final String RELATIVE_PATH = "/foo/bar/5MB.zip";
+    private static final String RELATIVE_PATH = "/foo/bar/";
 
     @Test
     public void returnsBatch_whenOptionalParametersAreNotSupplied() {
@@ -19,7 +19,7 @@ public class BatchBuilderTest {
                 .downloadFrom("net_address").apply()
                 .build();
 
-        BatchFile expectedBatchFile = new BatchFile("net_address", Optional.absent(), Optional.absent());
+        BatchFile expectedBatchFile = new BatchFile("net_address", Optional.absent(), "net_address");
         Batch expectedBatch = new Batch(DOWNLOAD_BATCH_ID, DOWNLOAD_BATCH_TITLE, Collections.singletonList(expectedBatchFile));
 
         assertThat(batch).isEqualTo(expectedBatch);
@@ -31,7 +31,7 @@ public class BatchBuilderTest {
                 .downloadFrom("net_address").withIdentifier(DOWNLOAD_FILE_ID).saveTo(RELATIVE_PATH, "5mb.zip").apply()
                 .build();
 
-        BatchFile expectedBatchFile = new BatchFile("net_address", Optional.of(DOWNLOAD_FILE_ID), Optional.of(RELATIVE_PATH));
+        BatchFile expectedBatchFile = new BatchFile("net_address", Optional.of(DOWNLOAD_FILE_ID), RELATIVE_PATH);
         Batch expectedBatch = new Batch(DOWNLOAD_BATCH_ID, DOWNLOAD_BATCH_TITLE, Collections.singletonList(expectedBatchFile));
 
         assertThat(batch).isEqualTo(expectedBatch);
